# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-chain merchant payment gateway API.",
        "title": "Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/transaction/rescan": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Re-fetches a transaction from the selected blockchain and replays it through wallet matching, payment, ledger, and webhook processing.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Rescan transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API secret returned when the domain was created or rotated",
                        "name": "X-API-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unix timestamp used in HMAC signature",
                        "name": "X-Gateway-Timestamp",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC-SHA256 over method + path/query + timestamp + raw body, optionally prefixed with sha256=",
                        "name": "X-Gateway-Signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "chain and tx_hash",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.V1ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.V1ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.V1ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/docs/integration-guide.md": {
            "get": {
                "description": "Returns the payment gateway integration guide as plain Markdown for developers and AI agents.",
                "produces": [
                    "text/markdown"
                ],
                "tags": [
                    "Documentation"
                ],
                "summary": "Integration guide",
                "responses": {
                    "200": {
                        "description": "Markdown integration guide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "integration guide not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/docs/money-event-catalog.md": {
            "get": {
                "description": "Returns the canonical and compatibility webhook event catalog as plain Markdown.",
                "produces": [
                    "text/markdown"
                ],
                "tags": [
                    "Documentation"
                ],
                "summary": "Money event catalog",
                "responses": {
                    "200": {
                        "description": "Markdown money event catalog",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "money event catalog not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/merchant/login": {
            "get": {
                "description": "Renders the hosted merchant portal login page with the OIDC sign-in action.",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Merchant Portal"
                ],
                "summary": "Show merchant login",
                "responses": {
                    "200": {
                        "description": "HTML merchant login page",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Authenticates a merchant with email and password, sets a merchant portal session cookie, and redirects to onboarding.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Merchant Portal"
                ],
                "summary": "Merchant email login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Merchant email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to merchant onboarding",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "HTML login page with validation error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "HTML login page with authentication error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.V1ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "X-API-Key header is required"
                },
                "result": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}